LayoutTests/imported/w3c/web-platform-tests/media-source/mediasource-sourcebuffer-mode-timestamps-expected.txt


FAIL audio/aac : If generate timestamps flag equals true and new mode equals "segments", then throw a TypeError exception and abort these steps. assert_throws_js: SourceBuffer#mode with generate timestamps flag true function "function () {
                                 sourceBuffer.mode = "segments";
                             }" threw object "InvalidAccessError: The object does not support the operation or argument." ("InvalidAccessError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL audio/mpeg : If generate timestamps flag equals true and new mode equals "segments", then throw a TypeError exception and abort these steps. assert_throws_js: SourceBuffer#mode with generate timestamps flag true function "function () {
                                 sourceBuffer.mode = "segments";
                             }" threw object "InvalidAccessError: The object does not support the operation or argument." ("InvalidAccessError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
